% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{jackknifeplus}
\alias{jackknifeplus}
\title{jackknifeplus}
\usage{
jackknifeplus(Xtrain, Ytrain, Xtest, Ytest, predfunction, a)
}
\arguments{
\item{Xtrain}{#' n training samples}

\item{Ytrain}{#' a response vector of size n}

\item{Xtest}{#' a test sample}

\item{Ytest}{#' a scalar, test sample's response value}

\item{predfunction}{#' a function that containing desired algorithm which input are Xtrain, Ytrain, Xtest, and output is predicted value(a scalar) of Xtest.}

\item{a}{#' probability coverage is 1-a}
}
\value{
#' conformal confidence interval of jackknife+
}
\description{
jackknifeplus
}
