% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackknifeplusCV_R_wrapper.R
\name{jackknifeplusCV_c_wrapper}
\alias{jackknifeplusCV_c_wrapper}
\title{jackknifeplusCV_c_wrapper}
\usage{
jackknifeplusCV_c_wrapper(Xtrain, Ytrain, Xtest, a, K)
}
\arguments{
\item{Xtrain}{#' n training samples, whose format must be matrix n*p}

\item{Ytrain}{#' a response matrix of dimension n * 1}

\item{Xtest}{#' a test sample as matrix, whose format must be consistent with Xtrain 1*p}

\item{a}{#' probability coverage is 1-2a 0<a<1}

\item{K}{#' divide data into K folders to do cross validation}
}
\value{
#' conformal confidence interval of jackknife+CV
}
\description{
jackknifeplusCV_c_wrapper
}
\examples{
Xtrain=matrix(rnorm(200),40,5)
Ytrain=matrix(rnorm(40),40,1)
Xtest=matrix(rnorm(5),1,5)
result = jackknifeplusCV_c_wrapper(Xtrain, Ytrain, Xtest, 0.05, 4)
}
