% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackknife_R_wrapper.R
\name{jackknife_c_wrapper}
\alias{jackknife_c_wrapper}
\title{jackknife_c_wrapper}
\usage{
jackknife_c_wrapper(Xtrain, Ytrain, Xtest, a)
}
\arguments{
\item{Xtrain}{#' n training samples, whose format must be consistent with input of predfunction user provides}

\item{Ytrain}{#' a response vector of size n}

\item{Xtest}{#' a test sample, whose format must be consistent with input of predfunction user provides}

\item{a}{#' probability coverage is 1-2a}
}
\value{
#' conformal confidence interval of jackknife+
}
\description{
jackknife_c_wrapper
}
\examples{
Xtrain=matrix(rnorm(200),40,5)
Ytrain=matrix(rnorm(40),40,1)
Xtest=matrix(rnorm(5),1,5)
result = jackknife_c_wrapper(Xtrain, Ytrain, Xtest, 0.05)
}
