% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{jackknifeplus}
\alias{jackknifeplus}
\title{jackknifeplus}
\usage{
jackknifeplus(Xtrain, Ytrain, Xtest, Ytest, predfunction, a)
}
\arguments{
\item{Xtrain}{#' n x p training data, 1st column should be 1s to account for intercept}

\item{Ytrain}{#' a vector of size n of class labels, from 0 to K-1}

\item{Xtest}{#' number of FIXED iterations of the algorithm, default value is 50}

\item{Ytest}{#' learning rate, default value is 0.1}

\item{predfunction}{#' (optional) initial starting values of beta for the algorithm, should be p x K matrix}

\item{a}{#' ridge parameter, default value is 0.1}
}
\value{
#'beta - p x K matrix of estimated beta values after numIter iterations; objective - (numIter + 1) length vector of objective values of the function that we are minimizing at each iteration (+ starting value)
}
\description{
jackknifeplus
}
