% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackknifeplus_R_wrapper.R
\name{jackknifeplus_c_wrapper}
\alias{jackknifeplus_c_wrapper}
\title{jackknifeplus_c_wrapper}
\usage{
jackknifeplus_c_wrapper(Xtrain, Ytrain, Xtest, a)
}
\arguments{
\item{Xtrain}{#' n training samples, whose format must be consistent with input of predfunction user provides}

\item{Ytrain}{#' a response vector of size n}

\item{Xtest}{#' a test sample, whose format must be consistent with input of predfunction user provides}

\item{a}{#' probability coverage is 1-2a}
}
\value{
#' conformal confidence interval of jackknife+
}
\description{
jackknifeplus_c_wrapper
}
\examples{
#' jackknifeplus_c_wrapper(Xtrain, Ytrain, Xtest, 0.05)

}
