% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackknifeplusCV_R_wrapper.R
\name{jackknifeplusCV_c_wrapper}
\alias{jackknifeplusCV_c_wrapper}
\title{jackknifeplusCV_c_wrapper}
\usage{
jackknifeplusCV_c_wrapper(Xtrain, Ytrain, Xtest, a, K)
}
\arguments{
\item{Xtrain}{#' n training samples, whose format must be consistent with input of predfunction user provides}

\item{Ytrain}{#' a response vector of size n}

\item{Xtest}{#' a test sample, whose format must be consistent with input of predfunction user provides}

\item{a}{#' probability coverage is 1-2a}

\item{K}{#' divide data into K folders to do cross validation}
}
\value{
#' conformal confidence interval of jackknife+
}
\description{
jackknifeplusCV_c_wrapper
}
\examples{
#' jackknifeplusCV_c_wrapper(Xtrain, Ytrain, Xtest, 0.05, 4)
}
